Improve tests a little
