--- test/fixtures/kill.mjs
@@ -1,5 +1,0 @@
-let p = nothrow($`sleep 9999`)
-setTimeout(() => {
-  p.kill()
-}, 100)
-await p
