+++ test/fixtures/markdown.md
@@ -1,0 +1,34 @@
+# Markdown
+
+ignore
+
+>
+> ```
+> echo ignore
+> ```
+
+```js
+await $`whoami`
+await $`echo ${__dirname}`
+```
+
+```js
+console.log(chalk.yellowBright(__filename))
+```
+
+```js
+await import('chalk')
+```
+
+```bash
+VAR=$(echo hello)
+echo "$VAR"
+```
+
+    console.log('world')
+
+Other code blocks are ignored:
+
+```css
+.ignore {}
+```
